update request around app
some back_end information change.

I've to update some request and code to keep displaying datas.

update everything new, app still functionnal.

add the feature for bind a mail to a folder. we can update the same mail too. Records in DB OK (update have to be fixed in backend, cause it make a second records for the same mail)